feat(reader): add reader subscribe to topic
