+++ src/lib/omerlo/reader/fetchers.ts
@@ -1,11 +1,11 @@
 import { accountsFetchers } from './endpoints/accounts';
-import { deviceFetchers } from './endpoints/device';
+import { notificationFetchers } from './endpoints/notification';
 import { oauthFetchers } from './endpoints/oauth';
 
 export const fetchers = (f: typeof fetch) => {
   return {
     ...accountsFetchers(f),
     ...oauthFetchers(f),
-    ...deviceFetchers(f),
+    ...notificationFetchers(f),
   };
 };
